Test signal handling by nodes
